# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Property {
  id: number;
  ownerId: number;
  title: string;
  description: string;
  type: string;
  location: string;
  city: string;
  pricePerNight: string;
  maxGuests: number;
  bedrooms: number;
  bathrooms: number;
  amenities: string[];
  images: string[];
  isVerified: boolean;
  isActive: boolean;
  /** @nullable */
  rating?: string | null;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface PropertyInput {
  ownerId: number;
  /** @minLength 1 */
  title: string;
  description: string;
  type: string;
  location: string;
  city: string;
  pricePerNight: string;
  maxGuests: number;
  bedrooms: number;
  bathrooms: number;
  amenities?: string[];
  images?: string[];
}

export interface PropertyUpdate {
  title?: string;
  description?: string;
  type?: string;
  location?: string;
  city?: string;
  pricePerNight?: string;
  maxGuests?: number;
  bedrooms?: number;
  bathrooms?: number;
  amenities?: string[];
  images?: string[];
  isVerified?: boolean;
  isActive?: boolean;
}

export interface Booking {
  id: number;
  propertyId: number;
  tenantId: number;
  /** @nullable */
  brokerId?: number | null;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalAmount: string;
  platformFee: string;
  ownerAmount: string;
  status: string;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BookingInput {
  propertyId: number;
  tenantId: number;
  /** @nullable */
  brokerId?: number | null;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalAmount: string;
  platformFee: string;
  ownerAmount: string;
  paymentMethod?: string;
  notes?: string;
}

export interface BookingStatusUpdate {
  status: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  nationalId?: string | null;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  nationalId?: string;
  role?: string;
  avatarUrl?: string;
}

export interface Review {
  id: number;
  bookingId: number;
  propertyId: number;
  reviewerId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  bookingId: number;
  propertyId: number;
  reviewerId: number;
  rating: number;
  comment?: string;
}

export interface AdminStats {
  totalProperties: number;
  totalBookings: number;
  totalUsers: number;
  totalRevenue: string;
  pendingBookings: number;
  activeBookings: number;
  completedBookings: number;
}

export interface SeedResult {
  message: string;
}

export type ListPropertiesParams = {
city?: string;
type?: string;
minPrice?: number;
maxPrice?: number;
guests?: number;
};

export type ListBookingsParams = {
tenantId?: number;
propertyId?: number;
status?: string;
};

export type ListReviewsParams = {
propertyId?: number;
};