# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SeaNight API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: properties
    description: Property operations
  - name: bookings
    description: Booking operations
  - name: users
    description: User operations
  - name: reviews
    description: Review operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: guests
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: tenantId
          in: query
          schema: { type: integer }
        - name: propertyId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a property
      parameters:
        - name: propertyId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/seed:
    post:
      operationId: seedData
      tags: [admin]
      summary: Seed demo data
      responses:
        "200":
          description: Seeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Property:
      type: object
      required: [id, ownerId, title, description, type, location, city, pricePerNight, maxGuests, bedrooms, bathrooms, amenities, images, isVerified, isActive, reviewCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
        location:
          type: string
        city:
          type: string
        pricePerNight:
          type: string
        maxGuests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        isActive:
          type: boolean
        rating:
          type: ["string", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    PropertyInput:
      type: object
      required: [ownerId, title, description, type, location, city, pricePerNight, maxGuests, bedrooms, bathrooms]
      properties:
        ownerId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
        location:
          type: string
        city:
          type: string
        pricePerNight:
          type: string
        maxGuests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string

    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        location:
          type: string
        city:
          type: string
        pricePerNight:
          type: string
        maxGuests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        isActive:
          type: boolean

    Booking:
      type: object
      required: [id, propertyId, tenantId, checkIn, checkOut, guests, totalAmount, platformFee, ownerAmount, status, paymentMethod, createdAt, updatedAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        tenantId:
          type: integer
        brokerId:
          type: ["integer", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        totalAmount:
          type: string
        platformFee:
          type: string
        ownerAmount:
          type: string
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    BookingInput:
      type: object
      required: [propertyId, tenantId, checkIn, checkOut, guests, totalAmount, platformFee, ownerAmount]
      properties:
        propertyId:
          type: integer
        tenantId:
          type: integer
        brokerId:
          type: ["integer", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        totalAmount:
          type: string
        platformFee:
          type: string
        ownerAmount:
          type: string
        paymentMethod:
          type: string
        notes:
          type: string

    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    User:
      type: object
      required: [id, name, email, role, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    UserInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        role:
          type: string
        avatarUrl:
          type: string

    Review:
      type: object
      required: [id, bookingId, propertyId, reviewerId, rating, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        propertyId:
          type: integer
        reviewerId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string

    ReviewInput:
      type: object
      required: [bookingId, propertyId, reviewerId, rating]
      properties:
        bookingId:
          type: integer
        propertyId:
          type: integer
        reviewerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string

    AdminStats:
      type: object
      required: [totalProperties, totalBookings, totalUsers, totalRevenue, pendingBookings, activeBookings, completedBookings]
      properties:
        totalProperties:
          type: integer
        totalBookings:
          type: integer
        totalUsers:
          type: integer
        totalRevenue:
          type: string
        pendingBookings:
          type: integer
        activeBookings:
          type: integer
        completedBookings:
          type: integer

    SeedResult:
      type: object
      required: [message]
      properties:
        message:
          type: string