# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalProperties: number;
  totalBookings: number;
  totalUsers: number;
  totalRevenue: string;
  pendingBookings: number;
  activeBookings: number;
  completedBookings: number;
}