# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  propertyId: number;
  tenantId: number;
  /** @nullable */
  brokerId?: number | null;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalAmount: string;
  platformFee: string;
  ownerAmount: string;
  paymentMethod?: string;
  notes?: string;
}