# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  ownerId: number;
  title: string;
  description: string;
  type: string;
  location: string;
  city: string;
  pricePerNight: string;
  maxGuests: number;
  bedrooms: number;
  bathrooms: number;
  amenities: string[];
  images: string[];
  isVerified: boolean;
  isActive: boolean;
  /** @nullable */
  rating?: string | null;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
}