# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  title?: string;
  description?: string;
  type?: string;
  location?: string;
  city?: string;
  pricePerNight?: string;
  maxGuests?: number;
  bedrooms?: number;
  bathrooms?: number;
  amenities?: string[];
  images?: string[];
  isVerified?: boolean;
  isActive?: boolean;
}