# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  bookingId: number;
  propertyId: number;
  reviewerId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}