# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  bookingId: number;
  propertyId: number;
  reviewerId: number;
  rating: number;
  comment?: string;
}