# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeaNight API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  nationalId?: string;
  role?: string;
  avatarUrl?: string;
}